+++ modules/linux-function/r-storage.tf
@@ -1,5 +1,5 @@
 module "storage" {
-  for_each = toset(var.storage_account_access_key == null ? ["enabled"] : [])
+  for_each = toset(var.use_existing_storage_account ? [] : ["enabled"])
 
   source  = "claranet/storage-account/azurerm"
   version = "7.3.0"
@@ -24,6 +24,7 @@ module "storage" {
   https_traffic_only_enabled         = var.storage_account_enable_https_traffic_only
   public_nested_items_allowed        = false
   advanced_threat_protection_enabled = var.storage_account_enable_advanced_threat_protection
+  shared_access_key_enabled          = !var.storage_uses_managed_identity
 
   # Identity
   identity_type = var.storage_account_identity_type
@@ -57,7 +58,7 @@ module "storage" {
 }
 
 resource "azurerm_storage_account_network_rules" "storage_network_rules" {
-  for_each = toset(var.storage_account_access_key == null && var.storage_account_network_rules_enabled ? ["enabled"] : [])
+  for_each = toset(!var.use_existing_storage_account && var.storage_account_network_rules_enabled ? ["enabled"] : [])
 
   storage_account_id = local.storage_account_output.id
 
@@ -75,8 +76,8 @@ resource "azurerm_storage_account_network_rules" "storage_network_rules" {
 }
 
 data "azurerm_storage_account" "storage" {
-  name                = local.storage_account_name
-  resource_group_name = var.resource_group_name
+  name                = var.use_existing_storage_account ? split("/", var.storage_account_id)[8] : local.storage_account_name
+  resource_group_name = var.use_existing_storage_account ? split("/", var.storage_account_id)[4] : var.resource_group_name
 
   depends_on = [module.storage]
 }
@@ -101,6 +102,8 @@ resource "azurerm_storage_blob" "package_blob" {
 }
 
 data "azurerm_storage_account_sas" "package_sas" {
+  for_each = toset(var.application_zip_package_path != null && !var.storage_uses_managed_identity ? ["enabled"] : [])
+
   connection_string = data.azurerm_storage_account.storage.primary_connection_string
   https_only        = false
   resource_types {