+++ modules/windows-function/locals.tf
@@ -3,7 +3,6 @@ locals {
 
   default_site_config = {
     always_on                              = !local.is_consumption
-    ip_restriction                         = concat(local.subnets, local.cidrs, local.service_tags)
     application_insights_connection_string = var.application_insights_enabled ? local.app_insights.connection_string : null
     application_insights_key               = var.application_insights_enabled ? local.app_insights.instrumentation_key : null
   }
@@ -103,8 +102,12 @@ locals {
     length(regexall("/3[12]$", cidr)) > 0 ? [cidrhost(cidr, 0), cidrhost(cidr, -1)] : [cidr]
   ]))
 
-  is_local_zip    = length(regexall("^(http(s)?|ftp)://", var.application_zip_package_path != null ? var.application_zip_package_path : 0)) == 0
-  zip_package_url = var.application_zip_package_path != null && local.is_local_zip ? format("%s%s&md5=%s", azurerm_storage_blob.package_blob[0].url, data.azurerm_storage_account_sas.package_sas.sas, filemd5(var.application_zip_package_path)) : var.application_zip_package_path
+  is_local_zip = length(regexall("^(http(s)?|ftp)://", var.application_zip_package_path != null ? var.application_zip_package_path : 0)) == 0
+  zip_package_url = (
+    var.application_zip_package_path != null && local.is_local_zip ?
+    format("%s%s&md5=%s", azurerm_storage_blob.package_blob[0].url, try(data.azurerm_storage_account_sas.package_sas["enabled"].sas, "?"), filemd5(var.application_zip_package_path)) :
+    var.application_zip_package_path
+  )
 
-  storage_account_output = var.storage_account_access_key == null ? module.storage["enabled"].storage_account_properties : null
+  storage_account_output = data.azurerm_storage_account.storage
 }